added checksum check and proper uninstall for exa

+++ exa/package.sh
@@ -1,22 +1,41 @@
 #!/bin/false "This script should be sourced in a shell, not executed directly"
 
-EXA_VERSION=0.10.1
 
 function dependencies() {
     echo ""
 }
 
 function install_package() {
-    local zipfile=exa-linux-x86_64-v${EXA_VERSION}.zip
-    curl -L https://github.com/ogham/exa/releases/download/v${EXA_VERSION}/${zipfile} --output /tmp/${zipfile}
-    unzip -o /tmp/${zipfile} -d ${DOTFILES_PREFIX}
+    # version info
+    local version=0.10.1
+    local zipfile=exa-linux-x86_64-v${version}.zip
+    local checksum=f0b321494944fee3d15a37b69f51a82f
+    local manifest=${DOTFILES_SHARE}/exa/manifest.txt
+
+    # grab archive
+    curl -L https://github.com/ogham/exa/releases/download/v${version}/${zipfile} --output /tmp/${zipfile}
+
+    # check archive
+    if [ "${checksum}" != $(md5sum /tmp/${zipfile} | awk '{print $1}') ]; then
+        >&2 echo Unexpected checksum
+        return 1
+    fi
+
+    # extract archive
+    mkdir -p $(dirname ${manifest})
+    unzip -o /tmp/${zipfile} -d ${DOTFILES_PREFIX} | tee ${manifest}
+    sed -i "s#\s*inflating:\ ##" ${manifest}
+    sed -i "/^Archive/d"         ${manifest}
+
+    # removing archive
     rm -rf /tmp/${zipfile}
 }
 
 function uninstall_package() {
-    rm -rvf ${DOTFILES_PREFIX}/man/{exa.*,exa_*} ${DOTFILES_PREFIX}/completions/exa.*
-    rm -rvf ${DOTFILES_BIN}/exa
-    unalias ls
+    set +x
+    for file in $(cat ${DOTFILES_SHARE}/exa/manifest.txt); do
+        rm -fv ${file}
+    done
 }
 
 function init_package() {
